requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.78.0
anyio==4.12.1
attrs==25.4.0
audioop-lts==0.2.1
cartesia==2.0.17
certifi==2026.1.4
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.128.3
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.11
iterators==0.2.0
jiter==0.13.0
multidict==6.7.1
notion-client==2.7.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
websockets==16.0
yarl==1.22.0
